Update UI Bedroom & Scheduler

+++ Assets/Script/Gameplay/Schedule/Event/UpdateStatusCurrencyEventSO.cs
@@ -4,19 +4,19 @@ using UnityEngine.Events;
 [CreateAssetMenu(fileName = "Update Status Currency EventSO", menuName = "Events/Update Status Currency EventSO")]
 public class UpdateStatusCurrencyEventSO : ScriptableObject
 {
-    public UnityAction<int> UpdateStatusCurrencyEvent;
+    public UnityAction<float> UpdateStatusCurrencyEvent;
 
-    public void OnRiase (int currenyValue)
+    public void OnRiase (float currenyValue)
     {
         UpdateStatusCurrencyEvent?.Invoke(currenyValue);
     }
 
-    public void Register(UnityAction<int> listener)
+    public void Register(UnityAction<float> listener)
     {
         UpdateStatusCurrencyEvent += listener;
     }
 
-    public void Unregister(UnityAction<int> listener)
+    public void Unregister(UnityAction<float> listener)
     {
         UpdateStatusCurrencyEvent -= listener;
     }
